Mechanical Analysis

Elements

Amaru.MechSolid — Type

MechSolid

A bulk finite element for mechanical equilibrium analyses.

Amaru.MechRod — Type

MechRod

A line finite element for mechanical equilibrium analyses.

Missing docstring.

Missing docstring for Amaru.MechRodSolidJoint. Check Documenter's build log for details.

Material models

Amaru.LinearElastic — Type

LinearElastic

A type for linear elastic materials.

Fields

E::Float64: Young Modulus

ν::Float64: Poisson ratio

Amaru.LinearElastic — Method

LinearElastic(; params...)

Creates an LinearElastic material type

Arguments

E: Young modulus

nu: Poisson ratio

Amaru.ElasticSolidState — Type

ElasticSolidState

A type for the state data of a LinearElastic type.

Fields

env::ModelEnv: Environment information

σ::Vec6: Stress tensor

ε::Vec6: Strain tensor

Amaru.DruckerPrager — Type

DruckerPrager

A type for linear elastic materials with Drucker Prager failure criterion.

Fields

E::Float64: Young Modulus

ν::Float64: Poisson ratio

α::Float64: Drucker Prager alpha paramter

κ::Float64: Drucker Prager kappa paramter

H::Float64: Hardening parameter

ρ::Float64: Density

Amaru.DruckerPrager — Method

DruckerPrager(; E, nu, alpha, kappa, H, rho)

Creates an DruckerPrager material type

Arguments

E: Young modulus

nu: Poisson ratio

alpha: Drucker Prager alpha parameter

kappa: Drucker Prager kappa parameter

H: Hardening parameter

rho: Density

Amaru.DruckerPragerState — Type

DruckerPragerState

A type for the state data of a DruckerPrager type.

Fields

env::ModelEnv: Environment information

σ::Vec6: Stress tensor

ε::Vec6: Strain tensor

εpa::Float64: Accumulated plastic strain

Δγ::Float64: Plastic multiplier

Amaru.VonMises — Type

VonMises

A type for linear elastic materials with Von Mises failure criterion.

Fields

E::Float64: Young modulus

ν::Float64: Poisson ratio

σy::Float64: Yielding stress

H::Float64: Hardening parameter

ρ::Float64: Density

Amaru.VonMises — Method

VonMises(; params...)

Creates an VonMises material type

Arguments

E: Young modulus

nu: Poisson ratio

fy: Yielding stress

H: Hardening parameter

Amaru.VonMisesState — Type

DruckerPragerState

A type for the state data of a DruckerPrager type.

Fields

env::ModelEnv: Environment information

σ::Vec6: Stress tensor

ε::Vec6: Strain tensor

εpa::Float64: Accumulated plastic strain

Δγ::Float64: Plastic multiplier

Amaru.ElasticRod — Type

ElasticRod

A type for linear elastic materials in rods.

Fields

E::Float64: Young Modulus

Amaru.ElasticRod — Method

ElasticRod(; params...)

Creates an ElasticRod material type

Arguments

E: Young modulus

rho: Density

Amaru.ElasticRodState — Type

ElasticRodState

A type for the state data of a ElasticRod type.

Fields

env::ModelEnv: environment information

σ::Float64: Axial stress

ε::Float64: Axial strain

Amaru.PPRod — Type

PPRod

A type for linear elastic perfectly plastic materials in rods.

Fields

E::Float64: Young modulus

fy::Float64: Yielding stress

H::Float64: Hardening parameter

Amaru.PPRod — Method

PPRod(; params...)

Creates an PPRod material type

Arguments

E: Young modulus

fy: Yielding stress

H: Hardening parameter

Amaru.PPRodState — Type

ElasticRodState

A type for the state data of a PPRod type.

Fields

env::ModelEnv

σ::Float64

ε::Float64

εp::Float64

Δγ::Float64