Mechanical Analysis

Elements

Amaru.MechSolid — Type

MechSolid

A bulk finite element for mechanical equilibrium analyses.

Amaru.MechRod — Type

MechRod

A line finite element for mechanical equilibrium analyses.

Amaru.MechRodSolidJoint — Type

MechRodSolidJoint

An interface element to link MechRod elements to bulk elements in mechanical equilibrium analyses.

Material models

Amaru.ElasticSolid — Type

ElasticSolid

A type for linear elastic materials.

Fields

E::Float64

Young Modulus

nu::Float64

Poisson ratio

ρ::Float64

Density

Amaru.ElasticSolid — Method

ElasticSolid(; E, nu, rho)

Creates an ElasticSolid material type

Arguments

E: Young modulus

nu: Poisson ratio

rho: Density

Amaru.ElasticSolidState — Type

ElasticSolidState

A type for the state data of a ElasticSolid type.

Fields

env::ModelEnv

Environment information

σ::Vector{Float64}

Stress tensor

ε::Vector{Float64}

Strain tensor

Amaru.DruckerPrager — Type

DruckerPrager

A type for linear elastic materials with Drucker Prager failure criterion.

Fields

E::Float64

Young Modulus

ν::Float64

Poisson ratio

α::Float64

Drucker Prager alpha paramter

κ::Float64

Drucker Prager kappa paramter

H::Float64

Hardening parameter

ρ::Float64

Density

Amaru.DruckerPrager — Method

DruckerPrager(; E, nu, alpha, kappa, H, rho)

Creates an DruckerPrager material type

Arguments

E: Young modulus

nu: Poisson ratio

alpha: Drucker Prager alpha parameter

kappa: Drucker Prager kappa parameter

H: Hardening parameter

rho: Density

Amaru.DruckerPragerState — Type

DruckerPragerState

A type for the state data of a DruckerPrager type.

Fields

env::ModelEnv

Environment information

σ::Vector{Float64}

Stress tensor

ε::Vector{Float64}

Strain tensor

εpa::Float64

Accumulated plastic strain

Δγ::Float64

Plastic multiplier

Amaru.VonMises — Type

VonMises

A type for linear elastic materials with Von Mises failure criterion.

Fields

E::Float64

Young modulus

ν::Float64

Poisson ratio

σy::Float64

Yielding stress

H::Float64

Hardening parameter

ρ::Float64

Density

Amaru.VonMises — Method

VonMises(; E, nu, fy, H, rho)

Creates an VonMises material type

Arguments

E: Young modulus

nu: Poisson ratio

fy: Yielding stress

H: Hardening parameter

rho: Density

Amaru.VonMisesState — Type

DruckerPragerState

A type for the state data of a DruckerPrager type.

Fields

env::ModelEnv

Environment information

σ::Vector{Float64}

Stress tensor

ε::Vector{Float64}

Strain tensor

εpa::Float64

Accumulated plastic strain

Δγ::Float64

Plastic multiplier

Amaru.ElasticRod — Type

ElasticRod

A type for linear elastic materials in rods.

Fields

E::Float64

Young Modulus

A::Float64

Section area

ρ::Float64

Density

Amaru.ElasticRod — Method

ElasticRod(; E, A, dm, rho)

Creates an ElasticRod material type

Arguments

E: Young modulus

A: Section area

dm: Diameter (only if A is not provided)

rho: Density

Amaru.ElasticRodState — Type

ElasticRodState

A type for the state data of a ElasticRod type.

Fields

env::ModelEnv

environment information

σ::Float64

Axial stress

ε::Float64

Axial strain

Amaru.PPRod — Type

PPRod

A type for linear elastic perfectly plastic materials in rods.

Fields

E::Float64

Young modulus

A::Float64

Section area

σy0::Float64

Yielding stress

H::Float64

Hardening parameter

ρ::Float64

Density

Amaru.PPRod — Method

PPRod(; E, A, fy, sig_y, H, rho, dm)

Creates an PPRod material type

Arguments

E: Young modulus

A: Section area

dm: Diameter (only if A is not provided)

fy: Yielding stress

H: Hardening parameter

rho: Density

Amaru.PPRodState — Type

ElasticRodState

A type for the state data of a PPRod type.

Fields

env::ModelEnv

σ::Float64

ε::Float64

εp::Float64

Δγ::Float64